Created in the View Model the method 'getProject' that searches for a project with an Id.
Added in 'ProjectDetailViewController' the property 'projectId' and updated the property 'project' to do some actions when it gets a value. Also added the property 'projectsViewModel'.
Created the method 'getProjectFromViewModel' that uses the new method in the View Model to search for a project, and updated the 'addTableView' method because now the Table View uses other methods to get the data.

Updated in 'MainViewController' the method 'rowTapped' and now it send the projectId instead of the project object.